requirements.txt
google-api-python-client
google-auth-oauthlib
python-docx
python-dotenv
openai
requests